Splitting project up into files
